updated documentation of network_opcodes
changed code to be compliant with documentation

+++ common/unit.hpp
@@ -5,15 +5,15 @@
 
 class unit
 {
-    sf::Uint16 x;
-    sf::Uint16 y;
+    sf::Uint8 x;
+    sf::Uint8 y;
 
 public:
     unit();
-    unit( sf::Uint16 x, sf::Uint16 y);
-    void set_position( sf::Uint16 x, sf::Uint16 y );
-    sf::Uint16 get_x();
-    sf::Uint16 get_y();
+    unit( sf::Uint8 x, sf::Uint8 y);
+    void set_position( sf::Uint8 x, sf::Uint8 y );
+    sf::Uint8 get_x();
+    sf::Uint8 get_y();
 
 };
